Implemented File System support (indexeddb)

+++ src/api.js
@@ -277,8 +277,13 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
     * You can't instantiate this class directly, you have to use a
     * {@link Database} object in order to create a statement.
     *
-    * **Warning**: When you close a database (using db.close()),
-    * all its statements are closed too and become unusable.
+    * **Warnings**
+    * 1. When you close a database (using db.close()), all
+    * its statements are closed too and become unusable.
+    * 1. After calling db.prepare() you must manually free the assigned memory
+    * by calling Statement.free(). Failure to do this will cause subsequent
+    * 'DROP TABLE ...' statements to fail with 'Uncaught Error: database table
+    * is locked'.
     *
     * Statements can't be created by the API user directly, only by
     * Database::prepare
@@ -815,8 +820,9 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
     * @memberof module:SqlJs
     * Open a new database either by creating a new one or opening an existing
     * one stored in the byte array passed in first argument
-    * @param {number[]} data An array of bytes representing
-    * an SQLite database file
+    * @param {number[]|string} data An array of bytes representing
+    * an SQLite database file or a path
+    * @param {Object} opts Options to specify a filename
     */
     function Database(name, data) {
     	this.filename = name;
@@ -824,6 +830,10 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
         if (data != null) {
             FS.createDataFile("/", this.filename, data, true, true);
         }
+
+        if(name === ":mem:")
+        	this.memoryFile = true;
+
         this.handleError(sqlite3_open(this.filename, apiTemp));
         this.db = getValue(apiTemp, "i32");
         registerFullTextSearch(this.db);
@@ -1083,6 +1093,7 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
         var binaryDb = FS.readFile(this.filename, { encoding: "binary" });
         this.handleError(sqlite3_open(this.filename, apiTemp));
         this.db = getValue(apiTemp, "i32");
+        registerFullTextSearch(this.db);
         return binaryDb;
     };
 
@@ -1096,7 +1107,7 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
     * Databases **must** be closed when you're finished with them, or the
     * memory consumption will grow forever
      */
-    Database.prototype["close"] = function close() {
+    Database.prototype["close"] = function close(unlink) {
         // do nothing if db is null or already closed
         if (this.db === null) {
             return;
@@ -1107,7 +1118,10 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
         Object.values(this.functions).forEach(removeFunction);
         this.functions = {};
         this.handleError(sqlite3_close_v2(this.db));
-        FS.unlink("/" + this.filename);
+
+        if(this.memoryFile || unlink) {
+        	FS.unlink("/" + this.filename);
+        }
         this.db = null;
     };
 
@@ -1212,7 +1226,7 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
             }
         }
         if (Object.prototype.hasOwnProperty.call(this.functions, name)) {
-            removeFunction(this.functions[name]);uuu
+            removeFunction(this.functions[name]);
             delete this.functions[name];
         }
         // The signature of the wrapped function is :
@@ -1292,4 +1306,49 @@ Module["onRuntimeInitialized"] = function onRuntimeInitialized() {
     // export Database to Module
     Module.Database = Database;
     Module.FullTextSearch = FullTextSearch;
+
+    // Because emscripten doesn't allow us to handle `ioctl`, we need
+    // to manually install lock/unlock methods. Unfortunately we need
+    // to keep track of a mapping of `sqlite_file*` pointers to filename
+    // so that we can tell our filesystem which files to lock/unlock
+    var sqliteFiles = new Map();
+
+    Module["register_for_idb"] = (customFS) => {
+      var SQLITE_BUSY = 5;
+
+      function open(namePtr, file) {
+        var path = UTF8ToString(namePtr);
+        sqliteFiles.set(file, path);
+      }
+
+      function lock(file, lockType) {
+        var path = sqliteFiles.get(file);
+        var success = customFS.lock(path, lockType)
+        return success? 0 : SQLITE_BUSY;
+      }
+
+      function unlock(file,lockType) {
+        var path = sqliteFiles.get(file);
+        customFS.unlock(path, lockType)
+        return 0;
+      }
+
+      let lockPtr = addFunction(lock, 'iii');
+      let unlockPtr = addFunction(unlock, 'iii');
+      let openPtr = addFunction(open, 'vii');
+		Module["_register_for_idb"](lockPtr, unlockPtr, openPtr)
+    }
+
+    // TODO: This isn't called from anywhere yet. We need to
+    // somehow cleanup closed files from `sqliteFiles`
+    Module["cleanup_file"] = (path) => {
+      let filesInfo = [...sqliteFiles.entries()]
+      let fileInfo = filesInfo.find(f => f[1] === path);
+      sqliteFiles.delete(fileInfo[0])
+    }
+
+    Module["reset_filesystem"] = () => {
+      FS.root = null;
+      FS.staticInit();
+    }
 };
